backend/requirements.txt

fastapi
uvicorn[standard]
pydantic
motor
python-dotenv
bcrypt
PyJWT
python-jose[cryptography]
python-multipart